Updated app.py

+++ resources/item.py
@@ -1,5 +1,10 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required,get_jwt_claims,
+    jwt_optional,
+    get_jwt_identity,
+    fresh_jwt_required
+    )
 from models.item import ItemModel
 
 class Item(Resource):
@@ -15,13 +20,14 @@ class Item(Resource):
         help="Every item needs a store id."
     )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {'message': 'Item not found'}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
@@ -37,22 +43,27 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims=get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message":"Admin Privilege required."},401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-
-        return {'message': 'Item deleted'}
+            return {'message': 'Item deleted'}
+        return {'message': 'Item Not found'},404
 
     def put(self, name):
         data = Item.parser.parse_args()
 
         item = ItemModel.find_by_name(name)
 
-        if item is None:
-            item = ItemModel(name, **data)
-        else:
+        if item:
             item.price = data['price']
+        else:
+            item = ItemModel(name, **data)
 
         item.save_to_db()
 
@@ -60,5 +71,14 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [x.json() for x in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [x.json() for x in ItemModel.find_all()]
+
+        if user_id:
+            return {"items":items},20
+        return {
+            'items': [item['name'] for item in items],
+            "message":"More data available if you log in.",
+            },200
